New Changes (#7)
* added feedback functionality (#8)

added feedback functionality (#8)

* Working in windows and linux (#9)

* Working for windows and linux

* Working for windows and linux

* Working for windows and linux

* Checking Permission and giving status (#12)

* Checking Permission and giving status

* Checking Permission and giving status

* Checking Permission and giving status

* Checking Permission and giving status

* Improve profile Information (#13)

* Improve profile Information

* Improve profile Information

* Improve profile Information

* Admin Section Added (#14)

* Basic Changes

* admin

* add libraries

* new

* Modified Admin Section

* Admin Section Updated

* Updated Profile Section View (#16)

* Basic Changes

* admin

* Fixed Profile Section View

* Updated Profile Section Information

* Delete error.log

* Admin Section Error.log file uploading error fixed (#18)

* Basic Changes

* admin

* Fixed Profile Section View

* Updated Profile Section Information

* Error file Deleted

* Syncing

* Login tracking added (#19)

* Basic Changes

* admin

* Fixed Profile Section View

* Updated Profile Section Information

* Error file Deleted

* Syncing

* Login Tracker Added

* Updated Login Tracker

* Handled Error

* Handled the isOnline section

* token is now available to isOnline method

* Database updated

* Log in with database success

* Login Tracker Working correctly

* Access with Ip

* isLogged now check for token

* added Database check of isLoggedin

* chnage

+++ controllers/LoginController.php
@@ -41,20 +41,28 @@ class LoginController extends BaseController
         exit();
       }
 
-      $session = new Session;
-      $options = array();
+      if(User::isOnline(User::$id,self::$token,self::$address))
+      {
+        $session = new Session;
+        $options = array();
 
-      $options['username'] = self::$username;
-      $options['name'] = User::$name;
-      $options['uid'] = User::$id;
-      $options['ipaddress'] = self::$address;
-      $options['token'] = self::$token;
-      $options['email'] = User::$useremail;
-      $options['iitpConnect_user_state'] = 'logged_in';
+        $options['username'] = self::$username;
+        $options['name'] = User::$name;
+        $options['uid'] = User::$id;
+        $options['ipaddress'] = self::$address;
+        $options['token'] = self::$token;
+        $options['email'] = User::$useremail;
+        $options['iitpConnect_user_state'] = 'logged_in';
 
-      $session->set($options);
-      $result = array('response' => 'success', 'text' => User::$username, 'message' => 'login successfull');
-      echo json_encode($result);
+        $session->set($options);
+        $result = array('response' => 'success', 'text' => User::$username, 'message' => 'login successfull');
+        echo json_encode($result);
+      }
+      else
+      {
+        $result = array('response' => 'error', 'text' => User::$username, 'message' => 'Something Went Wrong During Login');
+        echo json_encode($result);
+      }
       exit();
     }
     else
@@ -97,9 +105,18 @@ class LoginController extends BaseController
   public function UserLogout()
   {
     $session = new Session;
-    $session->destroy();
-    $result = array('response' => 'success', 'text' => 'Logout successfull');
-    echo json_encode($result);
+    
+    if(User::goOffline())
+    {
+      $session->destroy();
+      $result = array('response' => 'success', 'text' => 'Logout successfull');
+      echo json_encode($result);
+    }
+    else
+    {
+      $result = array('response' => 'error', 'text' => 'Something went Wrong During Logout');
+      echo json_encode($result);
+    }
     exit();
   }
 }
